refactor(jangar): split projection foreclosure evidence collector
